docs: DatePicker design tab
