+++ .dumi/theme/slots/Content/index.tsx
@@ -219,6 +219,7 @@ const Content: React.FC<{ children: ReactNode }> = ({ children }) => {
               </Space>
             </Typography.Paragraph>
           ) : null}
+          {meta.frontmatter.description}
           {children}
           {meta.frontmatter.filename && (
             <ContributorsList
