+++ components/date-picker/demo/design/pick-date.tsx
@@ -1,0 +1,9 @@
+import type { FC } from 'react';
+import React from 'react';
+import { DatePicker } from 'antd';
+
+const { _InternalPanelDoNotUseOrYouWillBeFired: PureDatePicker } = DatePicker;
+
+const Demo: FC = () => <PureDatePicker />;
+
+export default Demo;
