+++ components/date-picker/index.$tab-design.zh-CN.md
@@ -1,0 +1,7 @@
+## 组件定义
+
+DatePicker 的本质是选择（输入）日期型数据。
+
+## 基础使用
+
+<code src="./demo/design/pick-date.tsx" description="用于具体日期的选择。用户仅需要输入非常具体的日期信息时使用。">选择某天</code>
