+++ components/date-picker/index.en-US.md
@@ -2,13 +2,12 @@
 category: Components
 group: Data Entry
 title: DatePicker
+description: To select or input a date.
 cover: https://mdn.alipayobjects.com/huamei_7uahnr/afts/img/A*xXA9TJ8BTioAAAAAAAAAAAAADrJ8AQ/original
 coverDark: https://mdn.alipayobjects.com/huamei_7uahnr/afts/img/A*3OpRQKcygo8AAAAAAAAAAAAADrJ8AQ/original
 demo:
   cols: 2
 ---
-
-To select or input a date.
 
 ## When To Use
